feat: add sort and search on HomeFragment
